% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{install_token}
\alias{install_token}
\title{配置OPENGWAS_JWT与GITHUB_TOKEN的token}
\usage{
install_token(OPENGWAS_JWT = " ", GITHUB_TOKEN = " ")
}
\arguments{
\item{OPENGWAS_JWT}{IEU open GWAS数据库申请的token，注册网址：https://api.opengwas.io/profile/}

\item{GITHUB_TOKEN}{github申请的token。}
}
\description{
配置OPENGWAS_JWT与GITHUB_TOKEN的token
}
\examples{
install_token()

}
